Removed LowLevelMouseHook - introduced MouseHookWrapper and Nuget dependency on EventHook project.
