+++ RetailCoder.VBE/packages.config
@@ -3,10 +3,12 @@
   <package id="Antlr4" version="4.3.0" targetFramework="net45" />
   <package id="Antlr4.Runtime" version="4.3.0" targetFramework="net45" />
   <package id="Castle.Core" version="3.2.0" targetFramework="net45" />
+  <package id="EventHook" version="1.4.26" targetFramework="net45" />
   <package id="Ninject" version="3.2.2.0" targetFramework="net45" />
   <package id="Ninject.Extensions.Conventions" version="3.2.0.0" targetFramework="net45" />
   <package id="Ninject.Extensions.Factory" version="3.2.1.0" targetFramework="net45" />
   <package id="Ninject.Extensions.NamedScope" version="3.2.0.0" targetFramework="net45" />
+  <package id="Nito.AsyncEx" version="3.0.1" targetFramework="net45" />
   <package id="NLog" version="4.0.1" targetFramework="net45" />
   <package id="NLog.Schema" version="4.0.1" targetFramework="net45" />
 </packages>
